Fix failure due to the new token allocation algorithm - CASSANDRA-13701.
CASSANDRA-13701 enabled the new token allocation algorithm by default, which makes nodes take longer to bootstrap.
This breaks some tests that will require more time to pass fully and require sleep time to be added to match the new startup times.
Since the new token allocation algorithm isn't compatible with the OrderPreservingPartitioner and the ByteOrderedPartitioner, it was disabled for tests involving them.